Add files via upload
This is the BIDS script for one subject 
I have had some trouble finding all the information to fill in, but I dont think that is a big deal, because it is not our experiement. 

It works using EEGLAB and you will need a json master toolbox, i added the github for dwonload into the first lines of the scrpit as a comment.