BIDSifies the raw data
This is a new version of BIDS_Change file that was already uploaded. This script doesn't is suitable for when your raw data is already in the .bdf format, and you just want to move it to the right folders. I also added the seperation of raw data and derivatives (results of pipelines) into the strucuture