fixed thresholding problem, works on BIDS
This is Marida's version of the fieldtrip preprocessing and averaging script where she has fixed the problem with artifacts thresholding. I additionally mad ethe script work with our BIDS structure and save the intermediate and final data sets. Also it's now a function and we can call it from our Main script
 main